refactor(frontends/basic): split lowering helpers
