+++ src/frontends/basic/LowerEmit.hpp
@@ -1,0 +1,143 @@
+// File: src/frontends/basic/LowerEmit.hpp
+// Purpose: Declares IR emission helpers and lowering routines for BASIC.
+// Key invariants: Control-flow labels remain deterministic via BlockNamer.
+// Ownership/Lifetime: Operates on Lowerer state without owning AST or module.
+// Links: docs/class-catalog.md
+#pragma once
+
+private:
+void collectVars(const Program &prog);
+void collectVars(const std::vector<const Stmt *> &stmts);
+void lowerFunctionDecl(const FunctionDecl &decl);
+void lowerSubDecl(const SubDecl &decl);
+/// @brief Stack-allocate parameters and seed local map.
+void materializeParams(const std::vector<Param> &params);
+/// @brief Reset procedure-level lowering state.
+void resetLoweringState();
+/// @brief Lower function body statements.
+/// @return True if body contained statements, requiring finalization.
+bool lowerFunctionBody(const FunctionDecl &decl, const std::function<Value()> &defaultRet);
+/// @brief Emit final return block after body lowering.
+void finalizeFunction(const std::function<Value()> &defaultRet);
+void lowerStmt(const Stmt &stmt);
+RVal lowerExpr(const Expr &expr);
+/// @brief Lower a variable reference expression.
+/// @param expr Variable expression node.
+/// @return Loaded value and its type.
+RVal lowerVarExpr(const VarExpr &expr);
+/// @brief Lower a unary expression (e.g. NOT).
+/// @param expr Unary expression node.
+/// @return Resulting value and type.
+RVal lowerUnaryExpr(const UnaryExpr &expr);
+/// @brief Lower a binary expression.
+/// @param expr Binary expression node.
+/// @return Resulting value and type.
+RVal lowerBinaryExpr(const BinaryExpr &expr);
+/// @brief Lower logical (`AND`/`OR`) expressions with short-circuiting.
+/// @param expr Binary expression node.
+/// @return Resulting value and type.
+RVal lowerLogicalBinary(const BinaryExpr &expr);
+/// @brief Lower integer division and modulo with divide-by-zero check.
+/// @param expr Binary expression node.
+/// @return Resulting value and type.
+RVal lowerDivOrMod(const BinaryExpr &expr);
+/// @brief Lower string concatenation and equality/inequality comparisons.
+/// @param expr Binary expression node.
+/// @param lhs Pre-lowered left-hand side.
+/// @param rhs Pre-lowered right-hand side.
+/// @return Resulting value and type.
+RVal lowerStringBinary(const BinaryExpr &expr, RVal lhs, RVal rhs);
+/// @brief Lower numeric arithmetic and comparisons.
+/// @param expr Binary expression node.
+/// @param lhs Pre-lowered left-hand side.
+/// @param rhs Pre-lowered right-hand side.
+/// @return Resulting value and type.
+RVal lowerNumericBinary(const BinaryExpr &expr, RVal lhs, RVal rhs);
+/// @brief Lower a built-in call expression.
+/// @param expr Built-in call expression node.
+/// @return Resulting value and type.
+RVal lowerBuiltinCall(const BuiltinCallExpr &expr);
+
+public:
+// Built-in helpers
+RVal lowerLen(const BuiltinCallExpr &expr);
+RVal lowerMid(const BuiltinCallExpr &expr);
+RVal lowerLeft(const BuiltinCallExpr &expr);
+RVal lowerRight(const BuiltinCallExpr &expr);
+RVal lowerStr(const BuiltinCallExpr &expr);
+RVal lowerVal(const BuiltinCallExpr &expr);
+RVal lowerInt(const BuiltinCallExpr &expr);
+RVal lowerInstr(const BuiltinCallExpr &expr);
+RVal lowerLtrim(const BuiltinCallExpr &expr);
+RVal lowerRtrim(const BuiltinCallExpr &expr);
+RVal lowerTrim(const BuiltinCallExpr &expr);
+RVal lowerUcase(const BuiltinCallExpr &expr);
+RVal lowerLcase(const BuiltinCallExpr &expr);
+RVal lowerChr(const BuiltinCallExpr &expr);
+RVal lowerAsc(const BuiltinCallExpr &expr);
+RVal lowerSqr(const BuiltinCallExpr &expr);
+RVal lowerAbs(const BuiltinCallExpr &expr);
+RVal lowerFloor(const BuiltinCallExpr &expr);
+RVal lowerCeil(const BuiltinCallExpr &expr);
+RVal lowerSin(const BuiltinCallExpr &expr);
+RVal lowerCos(const BuiltinCallExpr &expr);
+RVal lowerPow(const BuiltinCallExpr &expr);
+RVal lowerRnd(const BuiltinCallExpr &expr);
+
+private:
+// Shared argument helpers
+RVal lowerArg(const BuiltinCallExpr &c, size_t idx);
+RVal ensureI64(RVal v, il::support::SourceLoc loc);
+RVal ensureF64(RVal v, il::support::SourceLoc loc);
+
+void lowerLet(const LetStmt &stmt);
+void lowerPrint(const PrintStmt &stmt);
+/// @brief Emit blocks for an IF/ELSEIF chain.
+/// @param conds Number of conditions (IF + ELSEIFs).
+/// @return Indices for test/then blocks and ELSE/exit blocks.
+IfBlocks emitIfBlocks(size_t conds);
+/// @brief Evaluate @p cond and branch to @p thenBlk or @p falseBlk.
+void lowerIfCondition(const Expr &cond,
+                      BasicBlock *testBlk,
+                      BasicBlock *thenBlk,
+                      BasicBlock *falseBlk,
+                      il::support::SourceLoc loc);
+/// @brief Lower a THEN/ELSE branch and link to exit.
+/// @return True if branch falls through to @p exitBlk.
+bool lowerIfBranch(const Stmt *stmt,
+                   BasicBlock *thenBlk,
+                   BasicBlock *exitBlk,
+                   il::support::SourceLoc loc);
+void lowerIf(const IfStmt &stmt);
+void lowerWhile(const WhileStmt &stmt);
+void lowerFor(const ForStmt &stmt);
+void lowerForConstStep(const ForStmt &stmt, Value slot, RVal end, RVal step, int64_t stepConst);
+void lowerForVarStep(const ForStmt &stmt, Value slot, RVal end, RVal step);
+ForBlocks setupForBlocks(bool varStep);
+void emitForStep(Value slot, Value step);
+void lowerNext(const NextStmt &stmt);
+void lowerGoto(const GotoStmt &stmt);
+void lowerEnd(const EndStmt &stmt);
+void lowerInput(const InputStmt &stmt);
+void lowerDim(const DimStmt &stmt);
+void lowerRandomize(const RandomizeStmt &stmt);
+
+// helpers
+Value emitAlloca(int bytes);
+Value emitLoad(Type ty, Value addr);
+void emitStore(Type ty, Value addr, Value val);
+Value emitBinary(Opcode op, Type ty, Value lhs, Value rhs);
+/// @brief Emit unary instruction of @p op on @p val producing @p ty.
+Value emitUnary(Opcode op, Type ty, Value val);
+void emitBr(BasicBlock *target);
+void emitCBr(Value cond, BasicBlock *t, BasicBlock *f);
+Value emitCallRet(Type ty, const std::string &callee, const std::vector<Value> &args);
+void emitCall(const std::string &callee, const std::vector<Value> &args);
+Value emitConstStr(const std::string &globalName);
+void emitTrap();
+void emitRet(Value v);
+void emitRetVoid();
+std::string getStringLabel(const std::string &s);
+unsigned nextTempId();
+Value lowerArrayAddr(const ArrayExpr &expr);
+void emitProgram(const Program &prog);
